requirements-rag.txt
chromadb==0.6.3
sentence-transformers==3.4.1
PyPDF2==3.0.1
python-docx==1.1.2
httpx==0.28.1
